+++ steering/platformio.ini
@@ -1,13 +1,43 @@
-[env:genericSTM32F446RE]
-platform = ststm32@13.0.0
-board = genericSTM32F446RE
-; can specify hardcoded port using upload_port flag
-upload_protocol = blackmagic
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[platformio]
+default_envs = q23-steering
+
+[env]
+lib_deps = 
+	jchisholm204/cal@^2.1.2
+	#jchisholm204/qfsae-dependencies@^0.0.4
+	#C:\Users\Jacob\Documents\GitHub\QFSAE\Libraries\dependencies
+	#C:\Users\Jacob\Documents\GitHub\QFSAE\Libraries\dependencies
+	#C:\Users\Jacob\Documents\GitHub\QFSAE\Libraries\cal
+	jchisholm204/st-f4can@^1.0.7
+	https://github.com/qfsae/FT800-FT813
+	https://github.com/evert-arias/EasyButton
 framework = arduino
-; change microcontroller
+build_flags = -D EVE_NHD_43
+check_tool = cppcheck
+check_flags = cppcheck: --addon=cert.py --suppress=unusedFunction
+
+[env:q23-steering]
+platform = ststm32
+board = genericSTM32F446RE
 board_build.mcu = stm32f446re
-monitor_speed = 115200
-; change MCU frequency
 board_build.f_cpu = 180000000L
+monitor_speed = 115200
+#debug_tool = stlink
+upload_protocol = stlink
 build_flags = 
-    -I../cal/include
+	${env.build_flags}
+	-D EVE_CS=PA4
+	-D EVE_PDN=8
+	-D EVE_SPI=SPI1
+	-DENABLE_HWSERIAL2
+	-I include
