--- steering/test/README.md
@@ -1,16 +1,0 @@
-This directory is intended for PlatformIO Unit Testing and project tests.
-
-Unit Testing is a software testing method by which individual units of
-source code, sets of one or more MCU program modules together with associated
-control data, usage procedures, and operating procedures, are tested to
-determine whether they are fit for use. Unit testing finds problems early
-in the development cycle.
-
-More information about PlatformIO Unit Testing:
-- https://docs.platformio.org/page/plus/unit-testing.html
-
-Unit tests can be run using the following command: 
-
-```
-pio test -e genericSTM32F446RE --verbose
-```
