--- steering/tftInfo.md
@@ -1,1 +1,0 @@
-# Steering Wheel PCB
